requirements.txt
# Core libraries
numpy==1.26.0

# Gym + MuJoCo
gymnasium==0.29.1
mujoco==2.3.3

# Stable Baselines3 (PPO, etc.)
stable-baselines3==2.5.0

# For text overlays (PIL) and progress
Pillow==10.0.0
tqdm==4.67.1
rich==13.3.4

# For movie creation (single_env_video_callback)
moviepy==2.1.2
# For optional logging
tensorboard==2.13.0
